Upgrade to .NET 8 (#191)
I just now saw #186, I didn't realize before that there was an effort in
place to do this work. There are parts of that PR that look
good—@luisquintanilla how should we proceed?

I pulled in the changes from the PR that I liked. ~Let's add the CPM in
a separate PR, since versions weren't bumped~.

I've bumped the versions, and I'm updating this PR to utilize the CPM
feature since it does improve the `PacakgeReference` experience.

---------

Co-authored-by: XiaoYun Zhang <xiaoyuz@microsoft.com>